+++ python/docstrings.cpp
@@ -3301,7 +3301,7 @@ Args:
 Returns:
     `True` if any point is inside the polygon set, `False` otherwise.)!");
 
-PyDoc_STRVAR(read_gds_function_doc, R"!(read_gds(infile, unit=0, tolerance=1e-2, filter=None) -> gdstk.Library
+PyDoc_STRVAR(read_gds_function_doc, R"!(read_gds(infile, unit=0, tolerance=0, filter=None) -> gdstk.Library
 
 Import a library from a GDSII stream file.
 
@@ -3309,7 +3309,8 @@ Args:
     infile (str or pathlib.Path): Name of the input file.
     unit (number): If greater than zero, convert the imported geometry
       to the this unit.
-    tolerance (number): Default tolerance for loaded paths.
+    tolerance (number): Default tolerance for loaded paths.  If zero or
+      negative, the library rounding size is used (`precision / unit`).
     filter (iterable of tuples): If not ``None``, only shapes with
       layer and data type in the iterable are read.
 
@@ -3321,7 +3322,7 @@ Examples:
     >>> top_cells = library.top_level()
     >>> filtered_lib = gdstk.read_gds("layout.gds", filter={(0, 1)}))!");
 
-PyDoc_STRVAR(read_oas_function_doc, R"!(read_oas(infile, unit=0, tolerance=1e-2) -> gdstk.Library
+PyDoc_STRVAR(read_oas_function_doc, R"!(read_oas(infile, unit=0, tolerance=0) -> gdstk.Library
 
 Import a library from an OASIS stream file.
 
@@ -3329,7 +3330,9 @@ Args:
     infile (str or pathlib.Path): Name of the input file.
     unit (number): If greater than zero, convert the imported geometry
       to the this unit.
-    tolerance (number): Default tolerance for loaded paths.
+    tolerance (number): Default tolerance for loaded paths and round
+      shapes.  If zero or negative, the library rounding size is used
+      (`precision / unit`).
 
 Returns:
     The imported library.